Refactor/naming style

+++ source/Parser/MetaInfo.cpp
@@ -1,11 +1,12 @@
-#include "parser/meta_info.h"
+#include "Parser/MetaInfo.h"
 
 #include <iostream>
 #include <unordered_set>
 
-#include "utils/utils.h"
+#include "Utils/Utils.h"
 
 namespace Aternyx {
+
 void AstTree::DebugInfo() const {
   std::cout << "metaStructMap_:" << std::endl;
   for (const auto& [key, value] : metaStructMap_) {
@@ -16,6 +17,7 @@ void AstTree::DebugInfo() const {
     std::cout << "  " << name << std::endl;
   }
 }
+
 void MetaField::AddAttributes(const std::string& attributeStr) {
   auto splitStr = Aternyx::StringLib::Split(attributeStr, ",");
   for (const auto& str : splitStr) {
@@ -37,25 +39,13 @@ void AstTree::EmplaceBack(Aternyx::MetaStruct&& metaStruct) {
   metaStructList.emplace_back(metaStruct);
   if (!metaStruct.baseTypeName.empty()) {
     size_t baseIndex = metaStructMap_[metaStruct.baseTypeName];
-    metaStructList[baseIndex].derivedTypeIndex.push_back(index);
+    metaStructList[baseIndex].derivedTypeIndex.push_back(static_cast<uint32_t>(index));
   }
 }
 
 void AstTree::RegisterTypeName(const std::string& fullTypeName) {
   typeNameSet_.insert(fullTypeName);
 }
-
-// void GetTypeNameList(const std::string& type_name, std::unordered_set<std::string>& type_list) {
-//   size_t temp_index = type_name.find('<');
-//   if (temp_index != std::string::npos) {
-//     type_list.insert(type_name.substr(0, temp_index));
-//     size_t temp_end_index = type_name.find_last_of('>') - 1;
-//     std::string left_type_name = type_name.substr(temp_index + 1, temp_end_index - temp_index);  // offset and length
-//     GetTypeNameList(left_type_name, type_list);
-//   } else if (!type_name.empty()) {
-//     type_list.insert(type_name);
-//   }
-// }
 
 std::string AstTree::GetTypeName(const std::string& typeName) {
   std::string finalTypeName;
